carsee_processor/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
onnxruntime==1.20.2
Pillow==11.0.0
numpy>=1.24.0,<2.0.0
